+++ tests/cassettes/core/test_download_bad_subtitle.yaml
